AppArmor: report true if apparmor parser is found on the host
Currently, CRI-O uses the IsSupported() method to determine
whether AppArmor is enabled. However, IsSupported() checks
more than just the existence of the AppArmor parser — it also
considers whether the environment is rootless and whether
runc is enabled. This leads to incorrect reporting in
scenarios where AppArmor is enabled on the host but the
environment is rootless, causing AppArmor to be shown as
disabled.

This patch replaces the use of IsSupported() with a check
using FindAppArmorParserBinary(), which only verifies the
existence of the AppArmor parser binary. This provides a
more accurate indication of weather  AppArmor is enabled
from CRI-O's perspective.

Signed-off-by: Feruzjon Muyassarov <feruzjon.muyassarov@est.tech>